requirements.txt
slack-bolt==1.20.1
slack-sdk==3.33.4
anthropic==0.42.0
python-dotenv==1.0.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
lxml_html_clean>=0.4.0
praw>=7.7.0
